update to check login, price, checkout
This new update is checking over if the login in function works properly and takes us to the home page of the website. If it succeeds it will check if the price on the first item is correct and see if it can add to cart. once added to cart it will go over if the check out tap works correctly and takes you to the next step of the checkout.